refactor service to reduce complexity
